dev: split integration tests into multiple files 🔨
