+++ tests/package.json
@@ -34,10 +34,10 @@
     }
   },
   "scripts": {
-    "test-with-logs": "mocha --printlogs -r ts-node/register 'tests/**/test-*.ts'",
+    "test-with-logs": "mocha --printlogs -r ts-node/register --require 'tests/setup-tests.ts' 'tests/**/test-*.ts'",
     "setup-local-typescript-api": "rm -rf node_modules/@keep-starknet-strange/madara-api-augment && cd ../typescript-api && npm run build && cp -r build ../tests/node_modules/@keep-starknet-strange/madara-api-augment",
-    "test": "echo 'WARNING Make sure to use node >= 20 otherwise it will not run' && mocha --parallel -r ts-node/register 'tests/**/test-*.ts' -- -j 4",
-    "test-seq": "mocha -r ts-node/register 'tests/**/test-*.ts'",
+    "test": "echo 'WARNING Make sure to use node >= 20 otherwise it will not run' && mocha --parallel -r ts-node/register --require 'tests/setup-tests.ts' 'tests/**/test-*.ts' -- -j 4",
+    "test-seq": "mocha -r ts-node/register --require 'tests/setup-tests.ts' 'tests/**/test-*.ts'",
     "watch": "npm-watch",
     "build": "npx tsc",
     "build-clean": "rm -r node_modules && npm i && npm run build",
